refactor(attention): what a deal is worth is its own file
classify.go crossed the 500-line cap. The drifting-deal classifier is the one
that answers a question about MONEY — everything else on the queue ranks on a
clock, and this ranks an amount against what the pipeline itself carries — so
it and its three money helpers move together.
